guest WB, shared based plan
+lint
+ edit and delete LS plans
+format + lint
